requirements.txt
alabaster==0.7.12
atomicwrites==1.2.1
attrs==18.2.0
Babel==2.6.0
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
colorama==0.4.0
dataframe==0.2.1.3
dill==0.2.8.2
Django==2.1.2
django-bootstrap-form==3.4
django-widget-tweaks==1.4.3
docutils==0.14
future==0.16.0
google-api-core==1.5.2
google-auth==1.6.0
google-cloud-core==0.26.0
google-cloud-vision==0.34.0
google-gax==0.15.16
googleapis-common-protos==1.5.5
grpcio==1.16.0
html5lib==1.0.1
idna==2.7
imagesize==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
more-itertools==4.3.0
nose==1.3.7
numpy==1.15.4
opencv-python==3.4.3.18
packaging==18.0
pandas==0.23.4
Pillow==5.3.0
pluggy==0.8.0
ply==3.8
protobuf==3.6.1
py==1.7.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
Pygments==2.2.0
pyparsing==2.3.0
PyPDF2==1.26.0
pyreadline==2.1
pytest==3.10.1
python-barcode==0.9.0
python-dateutil==2.7.5
pytz==2018.5
reportlab==3.5.9
requests==2.20.1
rsa==4.0
scipy==1.1.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.8.2
sphinxcontrib-websupport==1.1.0
tabulate==0.8.2
tenacity==4.12.0
urllib3==1.24.1
webencodings==0.5.1
xhtml2pdf==0.2.3
